Centered title + move file styling to own file
