+++ src/routes/klageskjema/begrunnelse/attachments/preview.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useState } from 'react';
 import styled from 'styled-components/macro';
 import { deleteAttachment } from '../../../../api/api';
 import { Attachment, AttachmentFile, toFiles } from '../../../../klage/attachment';
-import { FileFlexItem, FlexCenteredOnMobile } from '../../../../styled-components/common';
+import { FileFlexItem, FlexCenteredOnMobile } from '../../../../styled-components/file-preview';
 import { matchMediaQueries } from '../../../../styled-components/media-queries';
 
 interface Props {
