+++ src/styled-components/common.tsx
@@ -1,5 +1,4 @@
 import styled from 'styled-components/macro';
-import FileFC from 'forhandsvisningsfil';
 import { Normaltekst } from 'nav-frontend-typografi';
 import { device } from './media-queries';
 
@@ -23,69 +22,6 @@ export const SpaceBetweenFlexListContainer = styled.div`
     display: flex;
     justify-content: space-between;
     flex-flow: row wrap;
-`;
-
-export const FlexWithSpacingContainer = styled.div`
-    display: flex;
-    flex-flow: row wrap;
-
-    > * {
-        margin-right: 10px;
-    }
-
-    @media ${device.mobileS} {
-        > *:last-child {
-            margin-right: 10px;
-        }
-    }
-    @media ${device.mobileL} {
-        > *:last-child {
-            margin-right: 0;
-        }
-    }
-`;
-
-export const FlexCenteredOnMobile = styled(FlexWithSpacingContainer)`
-    @media ${device.mobileS} {
-        justify-content: center;
-    }
-    @media ${device.mobileL} {
-        justify-content: flex-start;
-    }
-`;
-
-export const CenteredOnMobile = styled.div`
-    @media ${device.mobileS} {
-        text-align: center;
-    }
-    @media ${device.mobileL} {
-        text-align: initial;
-    }
-`;
-
-export const FileFlexItem = styled(FileFC)`
-    margin-right: 25px;
-    margin-bottom: 25px;
-
-    @media ${device.mobileS} {
-        &:last-child {
-            margin-right: 25px;
-        }
-    }
-    @media ${device.mobileL} {
-        &:last-child {
-            margin-right: 0px;
-        }
-    }
-
-    .deleteLink {
-        svg {
-            width: 20px;
-            > path {
-                stroke: #ba3a26;
-            }
-        }
-    }
 `;
 
 export const WrapNormaltekst = styled(Normaltekst)`
