+++ src/styled-components/section-title.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components/macro';
-import { Innholdstittel } from 'nav-frontend-typografi';
+import { Systemtittel } from 'nav-frontend-typografi';
 
-export const SectionTitle = styled(Innholdstittel)`
+export const SectionTitle = styled(Systemtittel)`
     && {
         margin-top: 0;
         margin-bottom: 32px;
